# fix: Remove the `VerificationKey` from `ProverInstance`