--- a/barretenberg/cpp/src/barretenberg/client_ivc/client_ivc.test.cpp
+++ b/barretenberg/cpp/src/barretenberg/client_ivc/client_ivc.test.cpp
@@ -105,18 +105,20 @@ class ClientIVCTests : public ::testing::Test {
  */
 TEST_F(ClientIVCTests, Full)
 {
+    using VerificationKey = Flavor::VerificationKey;
+
     ClientIVC ivc;
     // Initialize IVC with function circuit
     Builder function_circuit = create_mock_circuit(ivc);
     ivc.initialize(function_circuit);
 
-    auto function_vk = ivc.prover_fold_output.accumulator->verification_key;
+    auto function_vk = std::make_shared<VerificationKey>(ivc.prover_fold_output.accumulator->proving_key);
     auto foo_verifier_instance = std::make_shared<VerifierInstance>(function_vk);
     // Accumulate kernel circuit (first kernel mocked as simple circuit since no folding proofs yet)
     Builder kernel_circuit = create_mock_circuit(ivc);
     FoldProof kernel_fold_proof = ivc.accumulate(kernel_circuit);
     // This will have a different verification key because we added the recursive merge verification to the circuit
-    auto function_vk_with_merge = ivc.prover_instance->verification_key;
+    auto function_vk_with_merge = std::make_shared<VerificationKey>(ivc.prover_instance->proving_key);
     auto kernel_vk = function_vk_with_merge;
     auto intermediary_acc = update_accumulator_and_decide_native(
         ivc.prover_fold_output.accumulator, kernel_fold_proof, foo_verifier_instance, kernel_vk);
@@ -137,7 +139,7 @@ TEST_F(ClientIVCTests, Full)
         foo_verifier_instance = construct_mock_folding_kernel(
             kernel_circuit, kernel_fold_output, function_fold_output, foo_verifier_instance);
         FoldProof kernel_fold_proof = ivc.accumulate(kernel_circuit);
-        kernel_vk = ivc.prover_instance->verification_key;
+        kernel_vk = std::make_shared<VerificationKey>(ivc.prover_instance->proving_key);
 
         intermediary_acc = update_accumulator_and_decide_native(
             ivc.prover_fold_output.accumulator, kernel_fold_proof, intermediary_acc, kernel_vk);