--- a/barretenberg/cpp/src/barretenberg/stdlib/recursion/honk/verifier/merge_verifier.test.cpp
+++ b/barretenberg/cpp/src/barretenberg/stdlib/recursion/honk/verifier/merge_verifier.test.cpp
@@ -84,7 +84,8 @@ class RecursiveMergeVerifierTest : public testing::Test {
             GoblinUltraComposer composer;
             auto instance = composer.create_prover_instance(outer_circuit);
             auto prover = composer.create_prover(instance);
-            auto verifier = composer.create_verifier(instance->verification_key);
+            auto verifier_instance = composer.create_verifier_instance(instance);
+            auto verifier = composer.create_verifier(verifier_instance->verification_key);
             auto proof = prover.construct_proof();
             bool verified = verifier.verify_proof(proof);