--- a/barretenberg/cpp/src/barretenberg/stdlib/recursion/honk/verifier/verifier.test.cpp
+++ b/barretenberg/cpp/src/barretenberg/stdlib/recursion/honk/verifier/verifier.test.cpp
@@ -147,8 +147,7 @@ template <typename BuilderType> class RecursiveVerifierTest : public testing::Te
     }
 
     /**
-     * @brief Construct a recursive verification circuit for the proof of an inner circuit then call check_circuit on
-     it
+     * @brief Construct a recursive verification circuit for the proof of an inner circuit then call check_circuit on it
      *
      */
     static void test_recursive_verification()
@@ -206,8 +205,7 @@ template <typename BuilderType> class RecursiveVerifierTest : public testing::Te
 
     /**
      * @brief Construct a verifier circuit for a proof whose data has been tampered with. Expect failure
-     * TODO(bberg #656): For now we get a "bad" proof by arbitrarily tampering with bits in a valid proof. It would
-     be
+     * TODO(bberg #656): For now we get a "bad" proof by arbitrarily tampering with bits in a valid proof. It would be
      * much nicer to explicitly change meaningful components, e.g. such that one of the multilinear evaluations is
      * wrong. This is difficult now but should be straightforward if the proof is a struct.
      */