--- a/barretenberg/cpp/src/barretenberg/sumcheck/instance/prover_instance.hpp
+++ b/barretenberg/cpp/src/barretenberg/sumcheck/instance/prover_instance.hpp
@@ -36,7 +36,6 @@ template <class Flavor> class ProverInstance_ {
     // currently commitment_key needs to be here, and not accessed through the proving key, since sometimes the proving
     // key is null during protogalaxy proving (TODO(https://github.com/AztecProtocol/barretenberg/issues/881)?)
     std::shared_ptr<CommitmentKey> commitment_key;
-    std::shared_ptr<VerificationKey> verification_key;
 
     ProverPolynomials prover_polynomials;
     WitnessCommitments witness_commitments;