In PR #4639 I introduced a tiny regression after doing the final master merge: while the VerificationKey wasn't used anymore in the PG prover workflow, as the PR aimed, it remained computed (and retrieved in some places) from the prover instance which nullified the benefits of precomputing verification keys for benchmarks. This PR addresses this.

# Benchmark

### Before

```
ninja: no work to do.
Benchmarking lock created at ~/BENCHMARK_IN_PROGRESS.
client_ivc_bench                                                                                                    100%   16MB  55.4MB/s   00:00    
2024-03-04T11:23:39+00:00
Running ./client_ivc_bench
Run on (16 X 3606.39 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x8)
  L1 Instruction 32 KiB (x8)
  L2 Unified 1024 KiB (x8)
  L3 Unified 36608 KiB (x1)
Load Average: 0.13, 0.72, 0.40
----------------------------------------------------------------
Benchmark                      Time             CPU   Iterations
----------------------------------------------------------------
ClientIVCBench/Full/6      32647 ms        26672 ms            1
Benchmarking lock deleted.
```

### After 
```
client_ivc_bench                                                                                                    100%   16MB  52.0MB/s   00:00    
2024-03-04T11:19:11+00:00
Running ./client_ivc_bench
Run on (16 X 3596.99 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x8)
  L1 Instruction 32 KiB (x8)
  L2 Unified 1024 KiB (x8)
  L3 Unified 36608 KiB (x1)
Load Average: 0.00, 0.00, 0.00
----------------------------------------------------------------
Benchmark                      Time             CPU   Iterations
----------------------------------------------------------------
ClientIVCBench/Full/6      29206 ms        24549 ms            1
```